+++ CHANGELOG.md
@@ -22,5 +22,6 @@
 * [X] 使类似`TODO()`的方法能够显示正确的日志。
 * [X] 优化和完善`LetterCase`和`ReferenceCase`以及相关代码。 
 * [X] 完成基本的顶层README文档。
+* [X] 进一步完善`LetterCase`和`ReferenceCase`以及相关代码。
 * [ ] breeze-text 提取和补充代码。
 * [ ] 重构`DataLoader`。